풀리퀘스트

+++ src/services/user-service.js
@@ -1,7 +1,11 @@
 const bcrypt = require('bcrypt');
-const { hashedPassword } = require('../utils/hashing');
 const { userDAO } = require('../db/dao/user-dao');
-const { validateEmail, validatePassword } = require('../utils/validator');
+const { hashedPassword } = require('../utils/hashing');
+const {
+	validateEmail,
+	validatePassword,
+	validateNickname,
+} = require('../utils/validator');
 
 class UserService {
 	async getUserByEmail(userEmail) {
@@ -10,7 +14,7 @@ class UserService {
 			if (!user) {
 				throw new Error(`이메일이 ${userEmail}인 유저가 존재하지 않습니다.`);
 			}
-			return user;
+			return { userEmail, nickname: user.nickname };
 		} catch (err) {
 			throw new Error('유저 조회에 실패했습니다.');
 		}
@@ -46,11 +50,16 @@ class UserService {
 		// 이미 존재하는 이메일인지 검사
 		const existingUser = await userDAO.findUserByEmail(userEmail);
 		if (existingUser) {
-			throw new Error('이미 존재하는 이메일입니다.');
+			const error = new Error('이미 존재하는 이메일입니다.');
+			error.status = 409;
+			throw error;
 		}
 
 		//이메일 유효성 검사
 		validateEmail(userEmail);
+
+		//닉네임 유효성 검사
+		validateNickname(nickname);
 
 		//비밀번호 유효성 검사
 		validatePassword(password);
@@ -61,7 +70,7 @@ class UserService {
 				nickname,
 				password: hashedPwd,
 			});
-			return user;
+			return { userEmail: user.userEmail, nickname: user.nickname };
 		} catch (err) {
 			throw new Error('유저 생성에 실패했습니다.');
 		}
@@ -69,23 +78,33 @@ class UserService {
 
 	async updateUser(userEmail, updatedInfo) {
 		const { password, nickname } = updatedInfo;
-		const hashedPwd = await hashedPassword(password);
+		let modifiedData;
 
 		if (!nickname) {
 			throw new Error('닉네임이 빈 값입니다.');
 		}
+
 		if (!password) {
-			throw new Error('패스워드가 빈 값입니다.');
-		}
+			validateNickname(nickname);
+			modifiedData = {
+				nickname,
+			};
+		} else {
+			const hashedPwd = await hashedPassword(password);
 
-		//비밀번호 유효성 검사
-		validatePassword(password);
+			validatePassword(password);
+			validateNickname(nickname);
 
-		try {
-			const { modifiedCount } = await userDAO.updateUser(userEmail, {
+			modifiedData = {
 				password: hashedPwd,
 				nickname,
-			});
+			};
+		}
+		try {
+			const { modifiedCount } = await userDAO.updateUser(
+				userEmail,
+				modifiedData,
+			);
 			if (!modifiedCount) {
 				throw new Error(`이메일이 ${userEmail}인 유저가 존재하지 않습니다.`);
 			}
@@ -97,21 +116,27 @@ class UserService {
 
 	async deleteUser(userEmail, typedPassword) {
 		//비밀번호 검증
-		const hashedPwd = await hashedPassword(typedPassword);
-		await this.verifyPassword(typedPassword, hashedPwd);
-
 		try {
+			const user = await userDAO.findUserByEmail(userEmail);
+			if (!user) {
+				throw new Error(`이메일이 ${userEmail}인 유저가 존재하지 않습니다.`);
+			}
+			const isValidPassword = await userService.verifyPassword(
+				typedPassword,
+				user.password,
+			);
+			if (!isValidPassword) {
+				throw new Error('비밀번호가 일치하지 않습니다.');
+			}
 			const { deletedCount } = await userDAO.deleteUser({
 				userEmail,
 			});
-
 			if (!deletedCount) {
 				throw new Error(`이메일이 ${userEmail}인 유저가 존재하지 않습니다.`);
 			}
-
 			return;
 		} catch (err) {
-			throw new Error(`유저 삭제에 실패했습니다: ${err.message}`);
+			throw new Error(`유저 삭제에 실패했습니다.`);
 		}
 	}
